requirements.txt
luigi==3.5.0
setuptools<81
pandas==2.2.1
sentry_sdk==1.42.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
psycopg2-binary==2.9.9
ipykernel
